refactor: refactor download.sh to check for zip file and directory, only download if both do not exist.

+++ ALFM/download.sh
@@ -31,14 +31,16 @@ files=(
     pneumonia.zip
 )
 
-# Function to download a file if it does not exist
+# Function to download a file if it and the unzipped directory do not exist
 download_file() {
     local file=$1
-    if [[ ! -f $file ]]; then
+    local dir_name=$(basename "$file" .zip)
+
+    if [[ ! -f "$file" && ! -d "$dir_name" ]]; then
         echo "Downloading $file ..."
         wget "${base_url}${file}" -O "$file"
     else
-        echo "$file exists. Skipping download."
+        echo "$file or directory $dir_name exists. Skipping download."
     fi
 }
 
@@ -46,4 +48,3 @@ download_file() {
 for file in "${files[@]}"; do
     download_file "$file"
 done
-
